Keypad + backspace
KEY_* enumerations in the curses headers seem to be generated from an OS specific file.  On OSX, ^H seems to be the
typical backspace character, however nowadays terminals send the Linux ^? in order to be compatible with other machines
when SSHing. With keypad off, curses interpets both ^H and ^? as key code 127. With keypad on, the interpretted code
depends on the system specific header. For Justin, backspace (^? on Linux) was being correctly converted to
KEY_BACKSPACE (value 263) by curses, but we we're handling that value in ce_config. For me, backspace (normally ^H on
Mac, but ^? in my case) was unabled to be mapped, and defaulted to value 127 which we were handling in the ce_config as
the backspace operation.

We now look for the KEY_BACKSPACE value when interpreting keys, and I've changed the settings in my terminal to send
^H as backspace. Maybe we should also support the raw 127 value?